[Pal/Linux-SGX] make OCALLs return longs not ints
Return type of INLINE_SYSCALL is long (%rax), so retun values of all
OCALLs must also be longs.

+++ Pal/src/host/Linux-SGX/sgx_api.h
@@ -20,7 +20,7 @@
 #include "sgx_arch.h"
 #include "pal_error.h"
 
-int sgx_ocall (unsigned long code, void * ms);
+long sgx_ocall (unsigned long code, void * ms);
 
 bool sgx_is_completely_within_enclave (const void * addr, uint64_t size);
 bool sgx_is_completely_outside_enclave(const void * addr, uint64_t size);
